refactor(PR4-C): merge exampleparent/dependency/ statetracker into dependencies.go
